trainable veriosn for MRI only baseline
