+++ src/minidl/runner/__init__.py
@@ -1,6 +1,6 @@
 from .base_runner import BaseRunner
 from .epoch_based_runner import EpochBasedRunner
-from .hooks import CheckpointHook, EarlyStoppingHook, Hook, HookBuilder, HookRegistry, TensorboardLoggerHook
+from .hooks import CheckpointHook, EarlyStoppingHook, Hook, HookBuilder, HookRegistry, TensorboardLoggerHook, WandbLoggerHook
 from .runner_registry import RunnerBuilder, RunnerRegistry
 
 __all__ = [
@@ -10,6 +10,7 @@ __all__ = [
     "TensorboardLoggerHook",
     "CheckpointHook",
     "EarlyStoppingHook",
+    "WandbLoggerHook",
     "HookRegistry",
     "HookBuilder",
     "RunnerRegistry",
